We managed to bypass anticheat 19
The function has been fixed so that it generates a safehouse for a random name that doesn't exist, then changes the safehouse and gives it to the real player, and finally kicks the nonexistent player.

The order in which the code is executed has also been reversed: first, the safehouse is executed, and then the item is removed from the player. This prevents them from being charged before they can claim it.