Prevent multiple instances of a safe house
Only one shelter is allowed to be viewed at a time.
A tooltip is added, which we have to translate
The text of the safe house menu is customized
The message from the shelters is personalized
If there is a spawn, the player is informed.
The Spanish text is added